refactor(utils): move getDeepActiveElement function to utils and update imports
