+++ pages/content/src/utils/utils.ts
@@ -1,11 +1,4 @@
-import DOMPurify from 'dompurify';
 // utils.ts
-export function stripHtml(html: string): string {
-  const temp = document.createElement('div');
-  temp.innerHTML = html;
-  return temp.textContent || temp.innerText || '';
-}
-
 export function parseHtmlToText(html: string): string {
   const tempDiv = document.createElement('div');
   tempDiv.innerHTML = html;
@@ -56,43 +49,8 @@ export function parseHtmlToText(html: string): string {
   return raw.replace(/\n{3,}/g, '\n\n').trim();
 }
 
-// export function convertTemplate(template: string): { convertedHtml: string; initialData: Record<string, string> } {
-//   console.log('convertedHtml:', template);
-//   const parser = new DOMParser();
-//   const doc = parser.parseFromString(template, 'text/html');
-//   const initialData: Record<string, string> = {};
-//   // 找到所有帶有 data-type="formtext" 的元素
-//   const fields = doc.querySelectorAll('[data-type="formtext"]');
-//   fields.forEach((elem, index) => {
-//     // 使用 label 屬性作為欄位 key，如果沒有，則用 index 補充
-//     const key = elem.getAttribute('label') || `field_${index}`;
-//     const defaultValue = elem.getAttribute('defaultvalue') || '';
-//     initialData[key] = defaultValue;
-//     // 建立 input 元件
-//     const input = doc.createElement('input');
-//     input.type = 'text';
-//     // 用 label 屬性作為 placeholder
-//     input.placeholder = key;
-//     input.value = defaultValue;
-//     input.name = key;
-//     // 可加入 onChange 事件，但這邊我們會在 popup 裡統一綁定
-//     // 將原先的 span 替換掉
-//     elem.parentNode?.replaceChild(input, elem);
-//   });
-//   console.log(' doc.body.innerHTML data:', doc.body.innerHTML);
-//   // 回傳轉換後的 innerHTML 與初始資料
-//   return { convertedHtml: doc.body.innerHTML, initialData };
-// }
-
-export function parseHtml(content: string): HTMLElement | null {
-  const cleanHTML = DOMPurify.sanitize(`<div>${content}</div>`);
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(cleanHTML, 'text/html');
-  return doc.body.firstElementChild as HTMLElement | null;
-}
-
 // 僅記錄游標位置無法區分這是哪個輸入元素的位置 ，下次插入時無法確定應該將內容插入到哪個輸入框
-// ElementPath 可用來驗證目前聚焦的元素是否與記錄時相同
+// ElementPath 可用來驗證目前聚焦的元素是否與記錄時相同，假如輸入框內已經有一段文字，而你需要在特定位置插入新內容，這時候需要記錄並定位該輸入框的位置。
 export function generateElementPath(element: Element): string {
   const path: string[] = [];
   let current: Element | null = element;
@@ -125,8 +83,6 @@ export function generateElementPath(element: Element): string {
 
 /**
  * 檢查元素是否為可編輯元素
- * @param element 要檢查的元素
- * @returns 是否為可編輯元素
  */
 export function isEditableElement(element: Element | null): boolean {
   return !!(
